sinai_nexus_backend/requirements.txt
numpy<2
fastapi
uvicorn
python-dotenv
pandas
pyarrow
rapidfuzz
google-generativeai
python-multipart

# embedding stack (safe versions — DO NOT CHANGE)
sentence-transformers==2.2.2
transformers==4.30.2
tokenizers==0.13.3
huggingface-hub==0.14.1

scikit-learn
unstructured
pypdf
python-docx
supabase
